Add exercise pause and automatic pause

+++ matcher/matcher.py
@@ -3,7 +3,7 @@ import logging
 
 from django.conf import settings
 
-from data.models import Comparison, Submission
+from data.models import Comparison, Submission, Exercise
 from radar.config import named_function
 import gc
 
@@ -14,13 +14,13 @@ logger = logging.getLogger("radar.matcher")
 def match(a):
     """
     Matches the submission A against already matched submissions for this exercise.
-    
+
     """
     logger.info("Matching submission %s", a)
     if Submission.objects.get(pk=a.pk).tokens == None:
         logger.info("Submission is not tokenized.")
         return False
-    
+
     def safe_div(a, b):
         return a / b if b > 0 else 0.0
 
@@ -34,12 +34,12 @@ def match(a):
 
     # Clear any unfinished business.
     Comparison.objects.filter(submission_a=a).delete()
-    
+
     # Match against template.
     logger.debug("Match %s vs template", a.student.key)
     l = 0
     while l < len(a.tokens) and l < len(a.exercise.template_tokens) and a.tokens[l] == a.exercise.template_tokens[l]:
-        l += 1    
+        l += 1
     ms = f(a.tokens, top_marks(len(a.tokens), l),
            a.exercise.template_tokens, top_marks(len(a.exercise.template_tokens), l),
            a.exercise.minimum_match_tokens)
@@ -55,32 +55,42 @@ def match(a):
     # Match against previously matched submissions.
     marks_a = a.template_marks()
     count_a = a.calculate_authored_token_count()
-    for b in a.submissions_to_compare:
-        
-        # Force garbage collection.
-        gc.collect()
-        
-        logger.debug("Match %s and %s", a.student.key, b.student.key)
-        ms = f(a.tokens, marks_a, b.tokens, b.template_marks(), a.exercise.minimum_match_tokens)
-        s = safe_div(ms.token_count(), (count_a + b.authored_token_count) / 2)
-        if s > settings.MATCH_STORE_MIN_SIMILARITY:
-            c = Comparison(submission_a=a, submission_b=b, similarity=s, matches_json=ms.json())
-            c.save()
-        if s > b.max_similarity:
-            b.max_similarity = s
-            b.save()
-        if a.max_similarity is None or s > a.max_similarity:
-            a.max_similarity = s
-            a.save()
-    
-    tail = Comparison.objects.filter(submission_a=a).exclude(submission_b__isnull=True)\
-        .order_by("-similarity")[settings.MATCH_STORE_MAX_COUNT:].values_list("id", flat=True)
-    Comparison.objects.filter(pk__in=list(tail)).delete()
-    
+    if count_a > 0:
+        for b in a.submissions_to_compare:
+
+            # Force garbage collection.
+            gc.collect()
+
+            logger.debug("Match %s and %s", a.student.key, b.student.key)
+            ms = f(a.tokens, marks_a, b.tokens, b.template_marks(), a.exercise.minimum_match_tokens)
+            s = safe_div(ms.token_count(), (count_a + b.authored_token_count) / 2)
+            if s > settings.MATCH_STORE_MIN_SIMILARITY:
+                c = Comparison(submission_a=a, submission_b=b, similarity=s, matches_json=ms.json())
+                c.save()
+            if s > b.max_similarity:
+                b.max_similarity = s
+                b.save()
+            if a.max_similarity is None or s > a.max_similarity:
+                a.max_similarity = s
+                a.save()
+
+        tail = Comparison.objects.filter(submission_a=a).exclude(submission_b__isnull=True)\
+            .order_by("-similarity")[settings.MATCH_STORE_MAX_COUNT:].values_list("id", flat=True)
+        Comparison.objects.filter(pk__in=list(tail)).delete()
+
     if a.max_similarity is None:
         a.max_similarity = 0.0
     a.authored_token_count = count_a
     a.save()
+
+    # Automatically pause exercise on many identical submissions.
+    if a.max_similarity > settings.AUTO_PAUSE_SIMILARITY:
+        if Submission.objects.filter(exercise=a.exercise,
+                    max_similarity__gt=settings.AUTO_PAUSE_SIMILARITY)\
+                .count() >= settings.AUTO_PAUSE_COUNT:
+            Exercise.objects.filter(id=a.exercise).update(paused=True)
+            return False
+
     return True
 
 class TokenMatchSet():
